Add Tangram dependency using CocoaPods
Podfile contains the CocoaPods configuration and Podfile.lock stores information about the referenced Pods.

CocoaPods creates a 'Pods' Xcode project for dependencies and modifies the App's Xcode project to reference them. It also creates an Xcode Workspace to contain both projects.

Pods directory is currently *not* checked in because the binaries are inconveniently large.
https://guides.cocoapods.org/using/using-cocoapods.html#should-i-check-the-pods-directory-into-source-control